[not verified] move heading to components repo
