+++ projects/js-packages/components/components/heading/style.module.scss
@@ -5,8 +5,3 @@
 .headline-small {
 	margin-bottom: var( --spacing-base );
 }
-
-.title-medium,
-.title-small {
-	margin-bottom: var( --spacing-base );
-}
